Refactoring: moved mat_mult_* into their seperate folder
Additionally, renamed documentation group BasicMatMultTrans and
BasicMatMultTransKernels to MatMultTrans and MatMultTransKernels,
respectively.
