+++ src/MatrixFunctions/mat_mult_trans/kernels/plp_mat_mult_trans_i32s_rv32im.c
@@ -1,7 +1,7 @@
 /* =====================================================================
  * Project:      PULP DSP Library
- * Title:        plp_mat_mult_i16s_rv32im.c
- * Description:  16-bit matrix multiplication kernel for RV32IM
+ * Title:        plp_mat_mult_i32s_rv32im.c
+ * Description:  32-bit matrix nultiplication kernel for RV32IM
  *
  * $Date:        22. December 2019
  * $Revision:    V0
@@ -32,44 +32,16 @@
 
 
 /**
-  @ingroup BasicMatMultTrans
+  @ingroup MatMultTrans
  */
 
 /**
-  @defgroup BasicMatMultTransKernels Matrix Multiplication Kernels
-  Computes the product of two matrices, the second of which is transposed.
-
-  The Matrix Matrix Multiplication computes the product of two matrices with dimensions MxN and NxO, the second one is stored transposed in memory.
-  The first matrix is accessed row wise, the second column wise, all values form the first are multiplied with the values of the second and then sum of the result gives the value for the result matrix.
-  <pre>
-      pDst[i,k] = pSrcA[i*M]*pSrcB[k*N] + pSrcA[i*M+1]*pSrcB[k*N+1] + ... + pSrcA[i*M+N-1]*pSrcB[k*N+N-1]
-  </pre>
-  There are separate functions int8, int16, and int32 data types. For lower precision integers (int8, int16), functions exploiting SIMD instructions are provided.
-
-  The naming of the functions follows the following pattern (for example plp_dot_prod_i32s_rv32im):
-  <pre>
-      \<pulp\> _ \<function name\> _ \<data type\>\<precision\>\<method\>_\<isa extension\>, with
-
-      data type = {f, i, q} respectively for floats, integers, fixed points
-
-      precision = {32, 16, 8} bits
-
-      method = {s, v, p} meaning single (or scalar, i.e. not using packed SIMD), vectorized (i.e. using SIMD instructions), and parallel (for multicore parallel computing), respectively.
-
-      isa extension = rv32im, xpulpv2, etc. of which rv32im is the most general one.
-
-  </pre>
-
-
- */
-
-/**
-  @addtogroup BasicMatMultTransKernels
+  @addtogroup MatMultTransKernels
   @{
  */
 
 /**
-  @brief Matrix multiplication of 16-bit integer matrices kernel for RV32IM extension.
+  @brief Matrix multiplication of 32-bit integer matrices kernel for RV32IM extension.
   @param[in]  pSrcA     points to the first input matrix
   @param[in]  pSrcB     points to the second input matrix
   @param[in]  M         height of the first input matrix
@@ -84,9 +56,9 @@
 
 #ifdef BASIC_VERSION
 
-void plp_mat_mult_trans_i16s_rv32im(
-                              const int16_t * __restrict__ pSrcA,
-                              const int16_t * __restrict__ pSrcB,
+void plp_mat_mult_trans_i32s_rv32im(
+                              const int32_t * __restrict__ pSrcA,
+                              const int32_t * __restrict__ pSrcB,
                               uint32_t M,
                               uint32_t N,
                               uint32_t O,
@@ -109,9 +81,9 @@ void plp_mat_mult_trans_i16s_rv32im(
 
 #else
 
-void plp_mat_mult_trans_i16s_rv32im(
-                              const int16_t * __restrict__ pSrcA,
-                              const int16_t * __restrict__ pSrcB,
+void plp_mat_mult_trans_i32s_rv32im(
+                              const int32_t * __restrict__ pSrcA,
+                              const int32_t * __restrict__ pSrcB,
                               uint32_t M,
                               uint32_t N,
                               uint32_t O,
@@ -120,9 +92,9 @@ void plp_mat_mult_trans_i16s_rv32im(
         uint32_t i= 0; // loop counter
         uint32_t j= 0; // loop counter
         uint32_t k= 0; // loop counter
-
-        uint32_t mod = N & 0x3;
         
+        uint32_t mod = N & 0x3;
+
         if(mod == 3){
           for(i=0; i < M; i++){
             for(k=0; k < O; k++){
@@ -192,6 +164,7 @@ void plp_mat_mult_trans_i16s_rv32im(
         }
 
 }
+
 #endif
 
 // #undef BASIC_VERSION
