+++ src/MatrixFunctions/mat_mult_trans/plp_mat_mult_trans_i32_parallel.c
@@ -1,7 +1,7 @@
 /* =====================================================================
  * Project:      PULP DSP Library
- * Title:        plp_mat_mult_i16.c
- * Description:  16-bit integer matrix multiplication glue code
+ * Title:        plp_mat_mult_i32_parallel.c
+ * Description:  parallel 32-bit integer matrix multiplication glue code
  *
  * $Date:        22. December 2019
  * $Revision:    V0
@@ -36,40 +36,52 @@
  */
 
 /**
-  @addtogroup BasicMatMultTrans
+  @addtogroup MatMultTrans
   @{
  */
 
 /**
-  @brief Glue code for matrix mutliplication of 16-bit integer matrices.
+  @brief Glue code for parallel matrix mutliplication of 32-bit integer matrices.
   @param[in]  pSrcA     points to the first input matrix
   @param[in]  pSrcB     points to the second input matrix
   @param[in]  M         height of the first input matrix
   @param[in]  N         width of the first input matrix and hight of the second
   @param[in]  O         width of the second input matrix
+  @param[in]  nPE       Number of cores to use
   @param[out] pDstC     points to the output matrix
   @return        none
  */
 
-void plp_mat_mult_trans_i16(
-                         const int16_t * __restrict__ pSrcA,
-                         const int16_t * __restrict__ pSrcB,
+void plp_mat_mult_trans_i32_parallel(
+                         const int32_t * __restrict__ pSrcA,
+                         const int32_t * __restrict__ pSrcB,
                          uint32_t M,
                          uint32_t N,
                          uint32_t O,
+                         uint32_t nPE,
                          int32_t * __restrict__ pDstC){
   
   if (rt_cluster_id() == ARCHI_FC_CID){
-    plp_mat_mult_trans_i16s_rv32im(pSrcA, pSrcB, M, N, O, pDstC);
+    printf("parallel processing supported only for cluster side\n");
+    return;
   }
   else{
-    plp_mat_mult_trans_i16v_xpulpv2(pSrcA, pSrcB, M, N, O, pDstC);
+    plp_mat_mult_instance_i32 args = {
+      .pSrcA = pSrcA,
+      .pSrcB = pSrcB,
+      .M = M,
+      .N = N,
+      .O = O,
+      .nPE = nPE,
+      .pDstC = pDstC
+    };
+    rt_team_fork(nPE,plp_mat_mult_trans_i32p_xpulpv2, (void*) &args);
   }
 
 }
 
 /**
-  @} end of BasicMatMultTrans group
+  @} end of MatMultTrans group
  */
